feat: add the unpatched beans for operate, tasklist and broker
